requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
python-dotenv
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
openpyxl
pandas
reportlab
weasyprint
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
psycopg2-binary
sqlalchemy
werkzeug
wtforms
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
psycopg2-binary
sqlalchemy
werkzeug
wtforms
openpyxl
pandas
reportlab
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
psycopg2-binary
sqlalchemy
werkzeug
wtforms
openpyxl
pandas
reportlab
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
psycopg2-binary
sqlalchemy
werkzeug
wtforms
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
psycopg2-binary
sqlalchemy
werkzeug
wtforms
pytz
